Change **args to **kwargs.
`BraxAgent` has `self.args`; I left this as-is. I did rename a couple of local variables from `args` to `kwargs`, though, so they could be used in calls like `agent(workspace, **kwargs)`.

I also fixed a couple of bugs where `NotImplementedError` was supposed to be raised (but there was a typo) and added one missing import of `math`.